Crud minor

See merge request ghost-team/api/springulart!23

+++ src/main/java/ca/ghostteam/springulart/controller/user/UserManagementController.java
@@ -34,7 +34,7 @@ public class UserManagementController {
 
     @ApiResponse(code = 200, message = "Successfully retrieved a user")
     @GetMapping("/{userId}")
-    public UserDTO getUser(@PathVariable("userId") Integer userId) {
+    public UserDTO getUser(@PathVariable("userId") Long userId) {
         return userService
                 .findAllUsers()
                 .stream()
@@ -61,17 +61,18 @@ public class UserManagementController {
 
     @ApiResponse(code = 204, message = "Successfully Deleted a user")
     @DeleteMapping(path = "{userId}")
-    public void deleteUser(@PathVariable("userId") Integer userId){
+    public void deleteUser(@PathVariable("userId") Long userId){
         userService.deleteUserById(userId);
     }
 
 
     @ApiResponse(code = 200, message = "Successfully updated a user")
     @PutMapping(path = "{userId}")
-    public UserDTO updateUser(@PathVariable("userId") Integer userId, @RequestBody UserDTO userDTO) throws Exception {
+    public UserDTO updateUser(@PathVariable("userId") Long userId, @RequestBody UserDTO userDTO) throws Exception {
         return this.userService
                 .updateUser(userId, userDTO)
                 .orElseThrow(() -> new IllegalStateException(String.format("User with ID %s cannot found", userId)));
+
     }
 
 }
